+++ actionpack/CHANGELOG.md
@@ -1,3 +1,40 @@
+*   The url_for helpers now support a new option called `path_params`.
+    This is very useful in situations where you only want to add a required param that is part of the route's URL but for other route not append an extraneous query param.
+
+    Given the following router...
+    ```ruby
+    Rails.application.routes.draw do
+      scope ":account_id" do
+        get "dashboard" => "pages#dashboard", as: :dashboard
+        get "search/:term" => "search#search", as: :search
+      end
+      delete "signout" => "sessions#destroy", as: :signout
+    end
+    ```
+
+    And given the following `ApplicationController`
+    ```ruby
+      class ApplicationController < ActionController::Base
+        def default_url_options
+          { path_params: { account_id: "foo" } }
+        end
+      end
+    ```
+
+    The standard url_for helper and friends will now behave as follows:
+
+    ```ruby
+    dashboard_path # => /foo/dashboard
+    dashboard_path(account_id: "bar") # => /bar/dashboard
+
+    signout_path # => /signout
+    signout_path(account_id: "bar") # => /signout?account_id=bar
+    signout_path(account_id: "bar", path_params: { account_id: "baz" }) # => /signout?account_id=bar
+    search_path("quin") # => /foo/search/quin
+    ```
+
+    *Jason Meller, Jeremy Beker*
+
 *   Change `action_dispatch.show_exceptions` to one of `:all`, `:rescuable`, or
     `:none`. `:all` and `:none` behave the same as the previous `true` and
     `false` respectively. The new `:rescuable` option will only show exceptions